pyproject.toml
[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "preprocessing"
version = "0.1.0"
description = "Preprocessing utilities for linguistic data"
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [
    {include = "preprocessing"}
]

[tool.poetry.dependencies]
python = ">=3.11,<3.15"
dkpro-cassis = ">=0.10.1,<0.11.0"
spacy = ">=3.8.7,<4.0.0"
stanza = ">=1.7.0"
nltk = ">=3.8.0"

[tool.poetry.group.dev.dependencies]
pytest = ">=7.4.3,<8.0.0"
